upgrades : add case when id didn't match + update of updatedAt when we update a todo field

+++ src/todos.js
@@ -41,9 +41,16 @@ async function createTodo (ctx) {
 async function deleteTodo (ctx) {
     const id = ctx.request.params.id
 
-    const result = await getDB().collection("todos").deleteOne({"_id" : ObjectId(id)})
+    const maybe_todo = await getDB().collection("todos").findOne({"_id" : ObjectId(id)})
 
-    ctx.body = { id: result.insertedId }
+    if(maybe_todo === undefined || maybe_todo === null) {
+        ctx.status = 404
+        ctx.body = { errorMsg: "No todo task found for this id" }
+    }
+    else {
+        const result = await getDB().collection("todos").deleteOne({"_id" : ObjectId(id)})
+        ctx.body = { id: result.insertedId }
+    }
 }
 
 async function updateTodo (ctx) {
@@ -52,28 +59,34 @@ async function updateTodo (ctx) {
     const title = ctx.request.body.title
     const completed = ctx.request.body.completed
 
-    if ((title === null || title === undefined || title == "") & (completed === null || completed === undefined || completed == "")) {
+    const maybe_todo = await getDB().collection("todos").findOne({"_id" : ObjectId(id)})
+
+    if(maybe_todo === undefined || maybe_todo === null) {
+        ctx.status = 404
+        ctx.body = { errorMsg: "No todo task found for this id" }
+    }
+    else if ((title === null || title === undefined || title == "") & (completed === null || completed === undefined || completed == "")) {
         ctx.status = 422
         ctx.body = { errorMsg: "Must have at least one parameter updated" }
     }
     else if ((title !== null || title !== undefined || title != "") & (completed === null || completed === undefined || completed == "")) {
         const result = await getDB().collection("todos").updateOne(
             {"_id" : ObjectId(id)},
-            {$set: { "title" : title}}
-        )        
+            {$set: { "title" : title, "updatedAt" : new Date(Date.now()).toISOString()}}
+        )
         ctx.body = { id: result.insertedId }
     }
     else if ((title === null || title === undefined || title == "") & (completed !== null || completed !== undefined || completed != "")) {
         const result = await getDB().collection("todos").updateOne(
             {"_id" : ObjectId(id)},
-            {$set: { "completed" : Boolean(completed)}}
+            {$set: { "completed" : Boolean(completed), "updatedAt" : new Date(Date.now()).toISOString()}}
         )        
         ctx.body = { id: result.insertedId }
     }
     else {
         const result = await getDB().collection("todos").updateOne(
             {"_id" : ObjectId(id)},
-            {$set: { "title" : title, "completed" : Boolean(completed)}}
+            {$set: { "title" : title, "completed" : Boolean(completed), "updatedAt" : new Date(Date.now()).toISOString()}}
         )        
         ctx.body = { id: result.insertedId }
     }
